refacto:
check_lines.py just countain the main func now.

 * check_lines.py,
   src/process.py : Here.
